Change C++ examples base type from float to double
In order to get more consistent tests results with the Python version examples.

+++ examples/build_map.cpp
@@ -61,8 +61,8 @@ int main(int argc, char *argv[])
 {
 	validateArgs(argc, argv);
 
-	typedef PointMatcher<float> PM;
-	typedef PointMatcherIO<float> PMIO;
+	typedef PointMatcher<double> PM;
+	typedef PointMatcherIO<double> PMIO;
 	typedef PM::TransformationParameters TP;
 	typedef PM::DataPoints DP;
