+++ examples/icp_simple.cpp
@@ -51,7 +51,7 @@ int main(int argc, char *argv[])
 	bool isCSV = true;
 	validateArgs(argc, argv, isCSV);
 	
-	typedef PointMatcher<float> PM;
+	typedef PointMatcher<double> PM;
 	typedef PM::DataPoints DP;
 	
 	// Load point clouds
